requirements.txt
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
Django==5.1.7
djangorestframework==3.15.2
dlib==19.24.6
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==2.2.4
opencv-python==4.11.0.86
pillow==11.1.0
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
